#1169 DDC-3343 - moved tests to correct test class
